BCC++ 5.5.1で警告封じオプション -w- を外した場合のコンパイル警告の除去. * 使用されていない変数はコメントアウトではなく削除した. * 英語版や他のプラットフォームとの互換をとるためにどうしても必要な部分の警告は   除去できなかった. * 他のコンパイラでは別の警告が残っている, もしくは新たに警告されるかもしれない. * 警告を黙らせる必要も薄れたので, makefile.bccからは -w- を除去.

+++ src/readdib.c
@@ -72,7 +72,7 @@ static DWORD PASCAL lread(int fh, VOID FAR *pv, DWORD ul)
 		ul -= MAXREAD;
 		hp += MAXREAD;
 	}
-	if (_lread(fh, (LPSTR)hp, (WORD)ul) != (WORD)ul)
+	if (_lread(fh, (LPSTR)hp, (WORD)ul) != ul)
 		return 0;
 	return ulT;
 }
@@ -108,7 +108,7 @@ static HPALETTE PASCAL NEAR MakeDIBPalette(LPBITMAPINFOHEADER lpInfo)
 		lpRGB = (RGBQUAD FAR *)((LPSTR)lpInfo + lpInfo->biSize);
 
 		/* copy colors from the color table to the LogPalette structure */
-		for (i = 0; i < lpInfo->biClrUsed; i++, lpRGB++)
+		for (i = 0; i < (WORD)lpInfo->biClrUsed; i++, lpRGB++)
 		{
 			npPal->palPalEntry[i].peRed = lpRGB->rgbRed;
 			npPal->palPalEntry[i].peGreen = lpRGB->rgbGreen;
